+++ src/low/curve2251-sse/relic_fb_add_low.c
@@ -61,7 +61,7 @@ void fb_addn_low(dig_t *c, const dig_t *a, const dig_t *b) {
 	*(__m128i *)(c + 2) = XOR(*(__m128i*)(a + 2), *(__m128i*)(b + 2));
 }
 
-void fb_addd_low(dig_t *c, const dig_t *a, const dig_t *b, int size) {
+void fb_addd_low(dig_t *c, const dig_t *a, const dig_t *b, size_t size) {
 	if (size == 2 * RLC_FB_DIGS) {
 		*(__m128i *)c = XOR(*(__m128i*)(a), *(__m128i*)(b));
 		*(__m128i *)(c + 2) = XOR(*(__m128i*)(a + 2), *(__m128i*)(b + 2));
